+++ boardslot.h
@@ -23,7 +23,7 @@ class BoardSlot
 
 		char toChar(void) const;
 
-		int value(void) const;
+		float value(void) const;
 
 	private:
 		void scanMoves(std::list<Move> *vals,
